+++ NEWS.md
@@ -1,5 +1,7 @@
-# rebird 1.3.9005
+# rebird 1.3.9006
 
+- Added `ebirdtaxonomyversion()` which retrieves data on available taxonomy versions.
+The version is added as an attribute to `rebird::tax` (#131).
 - Added `ebirdchecklist()`, which lets you view individual checklists (thanks @RichardLitt and @Rafnuss, #108).
 - Made `ebirdregioncheck()`, `ebirdhotspot()`, `ebirdloc()` defunct. 
 - API tests now use `vcr`, which saves "cassettes" of the tests which can be played